+++ src/api/methods/saveAmount.js
@@ -1,0 +1,11 @@
+const saveAmount = (methods) =>
+  new Promise((resolve) => {
+    methods
+      .saveAmount()
+      .call()
+      .then((amount) => {
+        resolve(amount);
+      });
+  });
+
+export default saveAmount;