+++ src/containers/RoundDetails/Details.js
@@ -1,3 +1,5 @@
+/* eslint-disable react/self-closing-comp */
+/* eslint-disable jsx-a11y/control-has-associated-label */
 /* eslint-disable react/prop-types */
 /* eslint-disable react-hooks/exhaustive-deps */
 /* eslint-disable no-unused-vars */
@@ -25,38 +27,49 @@ function Details({ roundData, roundId }) {
       <InputLabel
         label="Participantes"
         value={
-          <div className={styles.DetailParticipantsItem}>
+          <table className={styles.DetailParticipantsItem}>
+            <tr>
+              <th></th>
+              <th></th>
+              <th></th>
+              <th>Cobro</th>
+            </tr>
             {roundData.participantsData &&
               roundData.participantsData.map((participant) => (
-                <ul>
-                  <li>{participant.position}</li>
-                  <li>{formatAddress(participant.address)}</li>
-                  <li>{participant.admin && "Admin"}</li>
-                </ul>
+                <tr>
+                  <th>{participant.position}</th>
+                  <th>{formatAddress(participant.address)}</th>
+                  <th>{participant.admin && "Admin"}</th>
+                  <th>{participant.dateToWithdraw}</th>
+                </tr>
               ))}
-          </div>
+          </table>
         }
       />
-      <PageSubHeader title="Invitaciones" />
-      <Link
-        to={`/invitations?roundId=${roundId}`}
-        className={styles.RoundCardTitle}
-      >
-        <MailOutlined style={{ color: "white", fontSize: "20px" }} />
-      </Link>
-      <InputLabel
-        label="Invitaciones enviadas"
-        value={
-          <div className={styles.DetailParticipantsItem}>
-            {roundData.invitations &&
-              roundData.invitations.map((email) => (
-                <ul>
-                  <li>{email}</li>
-                </ul>
-              ))}
-          </div>
-        }
-      />
+      {roundData.stage === "ON_REGISTER_STAGE" && (
+        <>
+          <PageSubHeader title="Invitaciones" />
+          <Link
+            to={`/invitations?roundId=${roundId}`}
+            className={styles.RoundCardTitle}
+          >
+            <MailOutlined style={{ color: "white", fontSize: "20px" }} />
+          </Link>
+          <InputLabel
+            label="Invitaciones enviadas"
+            value={
+              <div className={styles.DetailParticipantsItem}>
+                {roundData.invitations &&
+                  roundData.invitations.map((email) => (
+                    <ul>
+                      <li>{email}</li>
+                    </ul>
+                  ))}
+              </div>
+            }
+          />
+        </>
+      )}
     </>
   );
 }